feat(narwhal): integrate new behavior for hydrating a block before committing it
One thing to note here. With the traditional consensus gossip pattern, we
are submitting the txs as part of the block part set that is gossiped
between validators/nodes. However, with the advent of narwhal, we only ever
send meta data that represents the txs (i.e. Collections), and that metadata
has to be hydrated before it can be commited. Otherwise, we commit no txs...