chore(tendermint): refactor CLI to encapsulate global state
this is super helpful for running tests. ATM you can't run multiple
tests against a command, or multiple of the same commands to run multiple
executables. You're left calling out to the os.Exec to run the go installed
binary. This removes that burden allowing us to build as part of the test.