feat(consensus): provide strategy dependency for informing consensus how to communicate peer state
the crux of the matter here is that the strategy by which consensus communicates
is dictated by a number of things. For the implementation coupled to the
CListMempool, everything is done in large full batches of parsets and its
corresponding block proposal. With the advent of narwhal, we no longer wish
to communicate the entire partset (i.e. all the txs for the block), but just
a "meta only" partset, that informs the validators of the colletions included
in the proposal. Those collections form the consensus part set along with the
rest of the block header, evidence and last commit. When using narwhal,
consensus sends roughly 1 part, for the entire consensus round, and after
the consensus part set has been trasmitted, the block is hydrated from the
meta data that corresponds to the proposed block's txs.