fix(narwhal): add timeout for reap completion and other narwhal tweaks
we're seeing issues in testnets where a narwhal primary is taking a long
time to complete a request. We've seen over 1 hour in some cases :exploding_head:...
We're adding a means to add a max ceiling on how long Reap will take to complete
as well as some logging to identify the time the narwhal operations are taking.